feat: use LoRA on a context instead of on a model
