+++ llama/addon/globals/getGpuInfo.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "napi.h"
+
+Napi::Value getGpuVramInfo(const Napi::CallbackInfo& info);
+Napi::Value getGpuDeviceInfo(const Napi::CallbackInfo& info);
+Napi::Value getGpuType(const Napi::CallbackInfo& info);
